+++ internal/component/prometheus/operator/configgen/config_gen_servicemonitor.go
@@ -6,6 +6,7 @@ import (
 	"sort"
 	"strings"
 
+	"github.com/alecthomas/units"
 	promopv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	"github.com/prometheus-operator/prometheus-operator/pkg/namespacelabeler"
 	commonConfig "github.com/prometheus/common/config"
@@ -25,6 +26,12 @@ func (cg *ConfigGenerator) GenerateServiceMonitorConfig(m *promopv1.ServiceMonit
 	}
 
 	cfg.JobName = fmt.Sprintf("serviceMonitor/%s/%s/%d", m.Namespace, m.Name, i)
+	if !cg.AllowArbitraryFileAccess {
+		if field := ServiceMonitorEndpointArbitraryFileField(ep); field != "" {
+			return nil, fmt.Errorf("serviceMonitor %s/%s endpoint %d uses %s, which is disallowed because allow_arbitrary_file_access is false; use %s instead", m.Namespace, m.Name, i, field, arbitraryFileFieldReplacement(field))
+		}
+	}
+
 	cfg.HonorLabels = ep.HonorLabels
 	if ep.HonorTimestamps != nil {
 		cfg.HonorTimestamps = *ep.HonorTimestamps
@@ -331,6 +338,48 @@ func (cg *ConfigGenerator) GenerateServiceMonitorConfig(m *promopv1.ServiceMonit
 	cfg.LabelLimit = uint(defaultIfNil(m.Spec.LabelLimit, 0))
 	cfg.LabelNameLengthLimit = uint(defaultIfNil(m.Spec.LabelNameLengthLimit, 0))
 	cfg.LabelValueLengthLimit = uint(defaultIfNil(m.Spec.LabelValueLengthLimit, 0))
+	if m.Spec.BodySizeLimit != nil {
+		if cfg.BodySizeLimit, err = units.ParseBase2Bytes(string(*m.Spec.BodySizeLimit)); err != nil {
+			return nil, fmt.Errorf("parsing bodySizeLimit from serviceMonitor: %w", err)
+		}
+	}
 
 	return cfg, cfg.Validate(cg.ScrapeOptions.GlobalConfig())
+}
+
+// ServiceMonitorEndpointArbitraryFileField returns the first arbitrary file field set on a ServiceMonitor endpoint.
+func ServiceMonitorEndpointArbitraryFileField(ep promopv1.Endpoint) string {
+	if ep.BearerTokenFile != "" { //nolint:staticcheck
+		return "bearerTokenFile"
+	}
+	if ep.TLSConfig == nil {
+		return ""
+	}
+	if ep.TLSConfig.CAFile != "" {
+		return "tlsConfig.caFile"
+	}
+	if ep.TLSConfig.CertFile != "" {
+		return "tlsConfig.certFile"
+	}
+	if ep.TLSConfig.KeyFile != "" {
+		return "tlsConfig.keyFile"
+	}
+	return ""
+}
+
+// arbitraryFileFieldReplacement returns the ServiceMonitor field that should be used instead
+// of the given arbitrary file field returned by ServiceMonitorEndpointArbitraryFileField.
+func arbitraryFileFieldReplacement(field string) string {
+	switch field {
+	case "bearerTokenFile":
+		return "bearerTokenSecret or authorization"
+	case "tlsConfig.caFile":
+		return "tlsConfig.ca"
+	case "tlsConfig.certFile":
+		return "tlsConfig.cert"
+	case "tlsConfig.keyFile":
+		return "tlsConfig.keySecret"
+	default:
+		return "secret or authorization fields"
+	}
 }